Update Level1_IO_Print Decimal Precision.c

+++ C_LEVEL1_Input Output Statements/Level1_IO_Print Decimal Precision.c
@@ -1,11 +1,11 @@
 #include<stdio.h>
 int main()
 {
-    float n;
+    double n;
     
     
-    scanf("%f",&n);
-    printf("%.2f",n);
+    scanf("%lf",&n);
+    printf("%.2lf",n);
     
     return 0;
 }
